Move multiple state tests to pytest
